+++ env.h
@@ -4,7 +4,7 @@
 #include "strings.h"
 
 typedef struct Record{
-    uint64_t key;
+    uint32_t key;
     Data data;
     struct Record *next;
 } Record;
@@ -15,8 +15,8 @@ typedef struct Environment{
 } Environment;
 
 Environment env_new(Environment *parent);
-void env_put(uint64_t key, Data value, Environment *env);
-void env_implicit_put(uint64_t key, Data value, Environment *env);
-Data env_get(uint64_t key, Environment *env, uint8_t beSilent);
+void env_put(uint32_t key, Data value, Environment *env);
+void env_implicit_put(uint32_t key, Data value, Environment *env);
+Data env_get(uint32_t key, Environment *env, uint8_t beSilent);
 void env_free(Environment env);
