+++ heap.h
@@ -3,16 +3,16 @@
 #include <stdint.h>
 #include "values.h"
 
-uint64_t heap_add_int(int32_t val);
-uint64_t heap_add_float(double val);
-uint64_t heap_add_str(uint64_t val);
-uint64_t heap_add_identifer(uint64_t val);
-uint64_t heap_add_logical(int32_t val);
+uint32_t heap_add_int(int32_t val);
+uint32_t heap_add_float(double val);
+uint32_t heap_add_str(uint32_t val);
+uint32_t heap_add_identifer(uint32_t val);
+uint32_t heap_add_logical(int32_t val);
 
-Data heap_get_data(uint64_t address);
-int32_t heap_get_int(uint64_t address);
-double heap_get_float(uint64_t address);
-uint64_t heap_get_str(uint64_t address);
-int32_t heap_get_logical(uint64_t address);
+Data heap_get_data(uint32_t address);
+int32_t heap_get_int(uint32_t address);
+double heap_get_float(uint32_t address);
+uint32_t heap_get_str(uint32_t address);
+int32_t heap_get_logical(uint32_t address);
 
 void heap_free();
