+++ native.h
@@ -4,7 +4,7 @@
 #include "interpreter.h"
 #include "env.h"
 
-Data handle_native(uint64_t name, Environment *env);
+Data handle_native(uint32_t name, Environment *env);
 void register_native(Environment *env);
 void unload_all();
 #endif
